Adds the ability to intialize from outside a venv.
The initialization script failed if run indirectly. For example, a
virual environment in ./.venv, running ./.venv/bin/install-dot-kernel
would fail because it could not find jupyter. The patch looks for
evidence the calling path is within a virtual environment, and when
it is, it sets PATH and VIRTUAL_ENV environment variables before running
the script.